refactor: move training helpers to utility modules
